requirements.txt
alembic==1.0.8
Click==7.0
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.8
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==5.1
six==1.12.0
SQLAlchemy==1.3.2
Werkzeug==0.15.2
WTForms==2.2.1
